fix(code): ensure fidelity with spec data for combinablePromotions
This was itching at my brain and I'm honestly unexcited about the
result. As currently constructed, this is the first place we have
enough information to build a result in the right order.

1. `allCombinablePromotions` doesn't know about the promotion code we
want, so it can't do any reordering based on that information
2. `.filter` can't reorder elements, only include/exclude them

This leaves us needing to do something "extra" given the promotion
code we have in scope, and there just aren't a lot of "nice" extras
that'll get us there.

At a push, we could maybe make this more of a piece using `flatMap` to
combine the filter and map, but I don't know if that buys much in
terms of clarity.